[NEST-BE-AI-Moderation] Change to return list of reasons instead of first reason violation and fix bug where other categories did not match properly
