+++ src/main/java/com/nest/core/moderation_service/utils/SightEngineUtils.java
@@ -1,0 +1,75 @@
+package com.nest.core.moderation_service.utils;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+
+import org.springframework.util.LinkedMultiValueMap;
+import org.springframework.util.MultiValueMap;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.JsonMappingException;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+public class SightEngineUtils {
+
+    private static List<String> categories = Arrays.asList("personal", "link", "spam", "content-trade",
+            "money-transaction", "extremism", "violence");
+
+    public static MultiValueMap<String, String> createTextModerationRequest(String text, String apiUser, String apiSecret) {
+        MultiValueMap<String, String> requestBody = new LinkedMultiValueMap<>();
+        requestBody.add("text", text);
+        requestBody.add("mode", "rules");
+        requestBody.add("lang", "en,fr,de,nl");
+        requestBody.add("categories", "profanity,personal,link,spam,content-trade,money-transaction,extremism,violence");
+        requestBody.add("api_user", apiUser);
+        requestBody.add("api_secret", apiSecret);
+        return requestBody;
+    }
+
+    public static String getViolationReasons(String jsonString) throws JsonMappingException, JsonProcessingException {
+        ObjectMapper mapper = new ObjectMapper();
+        JsonNode root = mapper.readTree(jsonString);
+        List<String> reasons = getProfanityReasons(root);
+
+        for(String category : categories) {
+            List<String> categoryReasons = getReasonByCategory(root, category);
+            if (!categoryReasons.isEmpty()) {
+                reasons.addAll(categoryReasons);
+            }
+        }
+        return String.join(", ", reasons);
+    }
+
+    // Profanity has different JSON structure than the other categories
+    private static List<String> getProfanityReasons(JsonNode root) {
+        List<String> reasons = new ArrayList<>();
+        JsonNode matches = root.path("profanity").path("matches");
+
+        if (matches.isArray() && matches.size() > 0) {
+            for (JsonNode match : matches) {
+                String type = match.path("type").asText();
+                String intensity = match.path("intensity").asText();
+
+                if (intensity.equalsIgnoreCase("high")) {
+                    reasons.add(type);
+                }
+            }
+        }
+        return reasons;
+    }
+
+    private static List<String> getReasonByCategory(JsonNode root, String category) {
+        List<String> reasons = new ArrayList<>();
+        JsonNode matches = root.path(category).path("matches");
+
+        if (matches.isArray() && matches.size() > 0) {
+            for (JsonNode match : matches) {
+                String type = match.path("type").asText();
+                reasons.add(type);
+            }
+        }
+        return reasons;
+    }
+}
